requirements (1).txt
streamlit
tensorflow
pandas
numpy
scikit-learn
Pillow
